ml_service/requirements.txt
paho-mqtt==1.6.1
scikit-learn==1.3.0
joblib==1.3.1
pandas==2.0.3
numpy==1.23.5
pyyaml==6.0.1
python-dotenv==1.0.0
influxdb-client==1.36.1
minio==7.1.15
pydantic
pydantic-settings
tensorflow==2.16.1
clickhouse_connect
fastapi
uvicorn